Create student app, refactor models
