+++ hakaton/student/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class StudentConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'student'
+    verbose_name = 'Студенты Практикума'
